requirements.txt
langchain
pandas
openai
pytest
python-dotenv
# sqlite3 is included in the Python standard library 